APIリクエストして、PRテーブルにcreateする処理を、モデルに移した
