+++ app/models/assigned_issue.rb
@@ -12,32 +12,33 @@ class AssignedIssue < ApplicationRecord
     end
 
     def api_request
-      issues = (1..1).map do |page|
-        client.search_issues('repo:fjordllc/bootcamp is:issue', options = { state: 'all', sort: 'updated', per_page: 10, page: page })
+      (1..1).map do |page|
+        client.search_issues('repo:fjordllc/bootcamp is:issue', { state: 'all', sort: 'updated', per_page: 10, page: page })
       end
     end
 
+    # rubocop:disable Metrics/MethodLength
     def create
       api_request.each do |issue|
         issue[:items].each do |i|
           next if AssignedIssue.exists?(number: i[:number])
+
           assigned_issue = AssignedIssue.new
           assigned_issue.number = i[:number]
+
           i[:labels].each do |label|
             next if label[:name].to_i.zero?
+
             assigned_issue.point = label[:name].to_i
           end
-          if i[:assignees].empty?
+          i[:assignees].each do |assignee|
             assigned_issue.assignees = []
-          else
-            i[:assignees].each do |assignee|
-              assigned_issue.assignees = []
-              assigned_issue.assignees << assignee[:id]
-            end
+            assigned_issue.assignees << assignee[:id]
           end
           assigned_issue.save!
         end
       end
     end
-end
+    # rubocop:enable Metrics/MethodLength
+  end
 end
